requirements/local.txt
-r base.txt

# development tools
ipdb==0.13.13

pre-commit==3.3.3
ruff==0.0.272
mypy==1.3.0

bandit==1.7.5
pip-audit==2.5.6
pipdeptree==2.9.3

httpx==0.24.1
pytest==7.3.2
pytest-asyncio==0.21.0
factory-boy==3.2.1
coverage==7.2.7